+++ lib/smart_app_launch/discovery_stu2_2_group.rb
@@ -1,7 +1,7 @@
 require_relative 'well_known_capabilities_stu2_test'
 require_relative 'cors_support_stu2_2_test'
 require_relative 'well_known_endpoint_stu2_2_test'
-require_relative 'url_helpers'
+require_relative 'cors_enababled_metadata_request_test'
 
 module SMARTAppLaunch
   class DiscoverySTU22Group < DiscoverySTU2Group
@@ -18,8 +18,8 @@ module SMARTAppLaunch
     well_known_index = children.find_index { |child| child.id.to_s.end_with? 'well_known_endpoint' }
     children[well_known_index] = children.pop
 
-    test from: :cors_support_stu2_2,
-         title: 'SMART well-known Endpoint Enables Cross-Origin Resource Sharing (CORS)',
+    test from: :smart_cors_support_stu2_2,
+         title: 'SMART .well-known/smart-configuration Endpoint Enables Cross-Origin Resource Sharing (CORS)',
          description: %(
                 For requests from any origin, CORS configuration permits access to the public discovery endpoints
                 (.well-known/smart-configuration and metadata). This test verifies that the
@@ -31,33 +31,6 @@ module SMARTAppLaunch
            }
          }
 
-    test do
-      include URLHelpers
-
-      title 'Conformance/CapabilityStatement Enables Cross-Origin Resource Sharing (CORS)'
-      description %(
-        For requests from any origin, CORS configuration permits access to the public discovery endpoints
-        (.well-known/smart-configuration and metadata). This test verifies that the metadata
-        request is returned with the appropriate CORS header.
-      )
-
-      input :url
-
-      fhir_client do
-        url :url
-        headers 'Origin' => Inferno::Application['inferno_host']
-      end
-
-      run do
-        fhir_get_capability_statement
-
-        assert_response_status(200)
-        inferno_origin = Inferno::Application['inferno_host']
-        cors_allow_origin = request.response_header('Access-Control-Allow-Origin')&.value
-        assert cors_allow_origin.present?, 'No `Access-Control-Allow-Origin` header received.'
-        assert cors_allow_origin == inferno_origin || cors_allow_origin == '*',
-               "`Access-Control-Allow-Origin` must be `#{inferno_origin}`, but received: `#{cors_allow_origin}`"
-      end
-    end
+    test from: :smart_cors_enabled_metadata_request
   end
 end
