+++ lib/smart_app_launch/openid_connect_group_stu2_2.rb
@@ -41,7 +41,7 @@ module SMARTAppLaunch
     fhir_user_claim_index = children.find_index { |child| child.id.to_s.end_with? 'fhir_user_claim' }
     children[fhir_user_claim_index] = children.pop
 
-    test from: :cors_support_stu2_2,
+    test from: :smart_cors_support_stu2_2,
          title: 'SMART FHIR User REST API Endpoint Enables Cross-Origin Resource Sharing (CORS)',
          description: %(
            For requests from a client's registered origin(s), CORS configuration permits access to the token endpoint
